+++ examples/02_stencil/stencil_parallel_4.cpp
@@ -216,13 +216,13 @@ int main(int argc, char* argv[])
 
     options_description desc_commandline;
     desc_commandline.add_options()
-        ("Nx", value<std::uint64_t>()->default_value(1024),
+        ("Nx", value<std::size_t>()->default_value(1024),
          "Elements in the x direction")
-        ("Ny", value<std::uint64_t>()->default_value(1024),
+        ("Ny", value<std::size_t>()->default_value(1024),
          "Elements in the y direction")
-        ("steps", value<std::uint64_t>()->default_value(100),
+        ("steps", value<std::size_t>()->default_value(100),
          "Number of steps to apply the stencil")
-        ("local-partitions", value<std::uint64_t>()->default_value(1),
+        ("local-partitions", value<std::size_t>()->default_value(1),
          "Number of local partitions on one locality")
         ("output", value<std::string>(),
          "Save output to file")
